Projects/StoryForge_agent/pyproject.toml
[project]
name = "storyforge-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "mcp[cli]>=1.25.0",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "tavily-python>=0.7.19",
]
